SCRD-1759 Specify cacert for keystone operations
(supersedes I96ab40380f5c59ed8c7e4665f69121bfbc79a23a)

Specify a default location for cacert in the keystone_v3
module so we don't have to include it in all the playbooks.
This will need to be changed to the default location on
SUSE when we remove hlinux support.

Change-Id: Ifee19e7ccccbbcaad883c903d5194db478466263