SCRD-2417 Remove keystone validate_config task
Upstream keystone-manage does not have a validate_config command. It was
written internally in response to a maxint validation bug that has since
been fixed upstream[1]. Oslo.config also validates config option types
and gives sensible responses when an option is set that doesn't make
sense, such as a string for an integer option. Remove the validate part
of the template task so that the keystone-reconfigure.yml can succeed.

[1] https://review.openstack.org/#/c/361406/

Change-Id: Ibecac071d99d9095e503717f9e7d14e9de253e17