SCRD-2611 Add domain observer role to policy
Modified policy json to allow user with
domain observer role and scoped to any project in the admin domain
to do a domain list. Also creates new domain observer role.

Notice that any project admin in the admin domain, which is the "Default"
domain out-of-the-box, can assign themselves with this role and go to town
with it. But give that this is the admin domain, at least we have some
restriction/protection against misuse.

Depends-on: I56c8715b03b1bffb2ee684c7ac007732bfa5ac7d
Upstream-ref: None
Change-Id: I671dc69231d5f08eb130c10811dc9cbdaedc068b