SCRD-2396 Fix shibboleth-www log permissions
On SUSE the shibboleth-sp package creates an additional log directory
/var/log/shibboleth-www that was not accounted for in the keystone
playbooks but was being managed in the logging playbooks. Since the
service creates the log files with root permissions, logrotate was
failing because it was trying to use the shibd user to manage the log
files. This patch ensures that the /var/log/shibboleth-www log files use
the same permissions as the /var/log/shibboleth log files so logrotate
will work smoothly.

Change-Id: I8d91d64454edd5ce4538765e69771bfbe97cbf9c