SCRD-2385 Fix apache site management on SUSE
Without this patch, the playbooks implement start and stop tasks for
Apache sites that depend on the a2ensite and a2dissite functionality,
with the intention of being able to enable and disable particular
services without stopping and starting Apache itself. In order to
replicate this functionality without using the a2ensite and a2dissite
utilities, the playbooks must instead add and remove the vhost file from
the main config directory.

Change-Id: Ied02b732d615572d5d2a97db1153de41df4a46e7