Revert "Do not create keystone user (bsc#1080169)" (bsc#1080169)
This reverts commit 6f594d6633c32c6be283669fefee06960195a82a.

As it turns out, the venv-openstack-keystone package does not actually
create the keystone user. It only uses the openstack-keystone package
for building the venv itself and then relies on the playbooks to create
users. The commit that this patch reverts was causing builds to fail due
to the missing user[1].

[1] https://ci.suse.de/job/ardana-job/748/console

Change-Id: I444961693d5542824bccb47e5ad10f194a2a1130