SCRD-3322 Fernet token distribution and rotation
This patch implements Keystone Fernet token setup, distribution,
and auto rotation that is closely resemble to the upstream
OpenStack Ansible logic.

The node with the Fernet rotation cron job is denoted the master node.
A monasca monitoring alarm definition is created for that node. If the
Fernet master node goes down, a CRITICAL alarm will be raised for the
cloud admin to take appropriate actions to either resurrect the master
node or designate a new master node if the existing master node is
permanently removed. Otherwise, skipping Fernet key rotation for a
prolong period of time may introduce security risks to the cloud.

This patch will also create a home dir for the keystone user and add
the plumbings (i.e. geneate ssh keys, setup authorized_keys, etc) in
order for it to perform Fernet key distribution. We also added
restrictions in authorized_keys to limit the ssh key for fernet
token distribution purpose only.

Because of configuring Keystone Fernet rotation require all the Keystone
API nodes in a cluster to be accounted for, the use of '--limit' option
in the Ansible play is not a wise idea. If the '--limit' option is used,
a message will be displayed back to the user indicating that Fernet
configuration will be skipped.

Because Fernet configuration requires all the controller nodes are
healthy and accounted for, Ferent auto key rotation should not be
performed while update is in progress. If manual update is to be
performed, we recommend stopping the Fernet auto rotation before
the update begins. To stop Fernet auto rotation, customer must run
the keystone-stop-fernet-auto-rotation.yml playbook.

Change-Id: I6941dd6ef3654b7a1490ec74cbf8173df6034704