do not change token provider on manual update (bsc#1097241)
In manual update situation where --limit is used, we do not want
to change the current token provider to Fernet as we will end up
with some controllers using Fernet token provider while others
still using UUID. This will be disastrous as tokens will be
randomly invalidated. This patch ensure we DO NOT support changing
token provider to Fernet until all the nodes are upgraded and ready as
Fernet required all the token-signing keys are in-sync for all the nodes
at all times.

This patch also isolates Fernet master node monitoring into a separate
playbook as we can't have it as part of keystone-configure-monasca.yml.
The reason is we need Monasca API service to be up and running in order
to create the notifications and alarm definitions. During fresh install,
monasca-deploy.yml is executed after keystone-deploy.yml which creates
a circular dependency situation. Instead, we'll have to update
ardana-deploy to include keystone-configure-fernet-monitoring.yml after
deploy-monasca.yml.

Change-Id: I81ec3bb981665dd924a9f0b13cd7b4470a92aacd