fudge host key validation (bsc#1097904)
Because we use rsync for fernet key distribution, and SSH host key checking
is enabled by default, we'll need to find a way to add the host keys
for the non-Fernet master nodes into the known_hosts of the Fernet master
node in an automated, non-interactive fashion. This is done by the clever
trick of runing a simple SSH command from the Fernet master node to the
given non-Fernet node using the "-o StrictHostKeyChecking=no" flag.
This will effectively add the host key for the first time and
*without* prompting the user.

Change-Id: I7ac551a87f642c3cec21f8edd71abaef582ffedb