properly add the host fingerprints (bsc#1097904)
This patch properly generate the host fingerprint by remotely
running a command with strict host key checking turned off.
For that to work, we'll need to first add the properly authorization
to the target host's authorized_keys file. Once the host fingerprint
is generated, we'll then replace the initial authorization with the
one that is for Fernet key rsync only.

This patch also updates the keystone-fernet-rotate.sh template to
use the IP address of the controller instead of the hostname as
the hostname is customizable and therefore may not be reliable.

In addition, this patch corrected a typo. It was supposed to be
'KEY_API', not 'KEY-API'.

Change-Id: I89f0197e601aa8598ec4091c76aa92e2dba9540a