Reruns of keystone credentials change (bsc#1102662)
This fix allows the ardana-keystone-credentials-change.yml playbook
to be rerun, and therefore idempotent. Yes, it allows multiple
successful runs of the playbook but this fix is especially
important when the playbook fails somewhere beyond this point. Without
this fix, subsequent attempts to run the playbook will now fail at this
point.

Change-Id: I2c15928b3a4aba7bfb555538afeeba0d5ac79c21