Updates keystone user if it exists (bsc#1102662)
Keystone users are created by the configuration tasks of each
service and these tasks are executed by the site.yml playbook. These
same configuration tasks are reused and reexecuted by the
ardana-keystone-credentials-change.yml playbook. When the keystone
module finds that the user already exists it immediately returns
resulting in credential change playbook failures when tasks attempt to
use the new passwords. This fix, when an existing user is found, updates
the users' properties like password, email, and description.

Change-Id: Ifc991f91016d78cd2da1f5da1b4a19a3d92f5d3f