SCRD-6077 add project_observer role to the admin user
In order to set a quota for a project, openstack CLI does
list_project first (don't ask me why) even if the project ID is
directly specify in the CLI argument. In order to list project, one
must have the project_observer role. Therefore, we need to grant it
to the keystone admin user for the service project as this is the
account we used to set quotas.

Change-Id: Ie2f842176f8a2b1b70b1ee9cb2db2b2035ff157d