fix domain_observer rule (SCRD-8720)
domain_observer rule is broken due to upstream policy refactoring
exercise in the Rockey release cycle.

With the removal of KeystoneToken from the token model, we longer have
the ability to use the token data syntax in the policy rules. This
change broke backward compatibility for those is deploying customized
Keystone policies. Unfortunately, we can't go back to KeystoneToken
model as the change was tightly coupled with the other refactored
authorization functionalities. Since the scope information is now
available in the credential dictionary, we can just make use of it
instead.

Also see https://review.openstack.org/#/c/629692

Change-Id: Ica2d38ff8969b250587e1372cb22780149cec920