support OpenID Connect WebSSO (SOC-10509)
This patch follows an established pattern of supporting WebSSO, namely
ADFS, via keystone-reconfiguration playbook. See

http://docserv.nue.suse.com/documents/SUSE_OpenStack_Cloud_9/
suse-openstack-cloud-operations/single-html/#id-1.5.7.13.6.3

Howver, unlike ADFS, we do not create the identity provider,
protocol, and mapping from the playbook. That's because of the
complexity of the mapping, which may take some iterations for
the customer to calibrate it correctly. Therefore, it's best
for the customer to perform these operations using openstack
CLI instead of having to stop/start the entire cluster for
Keystone reconfiguration.

Depends-On: Ib132d4de98d651a2c2e57fcf7a32fcc3b96e9200
Change-Id: Ie5c4dc1b951a874ad1d057de7f7e5664f61a1402